feat: improve use of lib and refactor code
