+++ package.json
@@ -8,6 +8,9 @@
     "type": "git",
     "url": "https://github.com/rodgeraraujo/uses-js.git"
   },
+  "bin": {
+    "use": "index.js"
+  },
   "dependencies": {
     "process": "^0.11.10"
   }
